+++ Classes/ARTiledImageScrollView.h
@@ -34,6 +34,9 @@
 /// Set a background image, displayed while tiles are being downloaded.
 @property (readwrite, nonatomic) NSURL *backgroundImageURL;
 
+/// Set a background image, displayed while tiles are being downloaded.
+@property (readwrite, nonatomic) UIImage *backgroundImage;
+
 /// Point on which to center the map by default, removed when panned.
 @property (nonatomic, assign) CGPoint centerPoint;