+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Next
+
+* Added `ARTiledImageScrollView.backgroundImage` - [@dblock](https://github.com/dblock).
+
 ## 1.0.0 (3/15/2013)
 
 * Initial public release - [@dblock](https://github.com/dblock).